// www/app/api/models/Participant.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from "../runtime";
/**
 *
 * @export
 * @interface Participant
 */
export interface Participant {
  /**
   *
   * @type {any}
   * @memberof Participant
   */
  id: any | null;
  /**
   *
   * @type {any}
   * @memberof Participant
   */
  speaker: any | null;
  /**
   *
   * @type {any}
   * @memberof Participant
   */
  name: any | null;
}

/**
 * Check if a given object implements the Participant interface.
 */
export function instanceOfParticipant(value: object): boolean {
  let isInstance = true;
  isInstance = isInstance && "id" in value;
  isInstance = isInstance && "speaker" in value;
  isInstance = isInstance && "name" in value;

  return isInstance;
}

export function ParticipantFromJSON(json: any): Participant {
  return ParticipantFromJSONTyped(json, false);
}

export function ParticipantFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): Participant {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    id: json["id"],
    speaker: json["speaker"],
    name: json["name"],
  };
}

export function ParticipantToJSON(value?: Participant | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    id: value.id,
    speaker: value.speaker,
    name: value.name,
  };
}